requirements.txt
Pillow
PyMuPDF
redis
crewai
crewai-tools
cryptography
flask
flask-cors
langchain-huggingface
langchain-openai
litellm
numpy
pdf2image
pytesseract
python-dotenv
qdrant-client
sentence-transformers
tenacity
tiktoken
waitress
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121